added validation to server

+++ server/routes/permissions.js
@@ -5,8 +5,14 @@ const { Permission } = require("../models");
 router.post("/", async (req, res) => {
   const { name } = req.body;
   try {
+    const isExistingPermission = await Permission.findOne({ where: { name } });
+    if (isExistingPermission) {
+      return res
+        .status(400)
+        .json({ message: "Permission with provided name already exists." });
+    }
     const permission = await Permission.create({ name });
-    return res.json(permission);
+    return res.status(201).json(permission);
   } catch (err) {
     console.log(err);
     return res.status(500).json(err);
@@ -27,6 +33,12 @@ router.put("/:id", async (req, res) => {
   const { id } = req.params;
   const { name } = req.body;
   try {
+    const isExistingPermission = await Permission.findOne({ where: { name } });
+    if (isExistingPermission) {
+      return res
+        .status(400)
+        .json({ message: "Permission with provided name already exists." });
+    }
     await Permission.update(
       { name },
       {
@@ -46,6 +58,12 @@ router.put("/:id", async (req, res) => {
 router.delete("/:name", async (req, res) => {
   const { name } = req.params;
   try {
+    const permission = await Permission.findOne({ where: name });
+    if (!permission) {
+      return res
+        .status(404)
+        .json({ message: "Permission with provided name does not exist." });
+    }
     await Permission.destroy({
       where: {
         name,
