+++ server/routes/roles.js
@@ -6,14 +6,26 @@ router.post("/", async (req, res) => {
   const { name, selectedPermissions } = req.body;
   const ids = selectedPermissions.map((permission) => permission.id);
   try {
+    const isExistingRole = await Role.findOne({ where: { name } });
+    if (isExistingRole) {
+      return res.status(400).json({
+        message: "Role with provided name already exists.",
+      });
+    }
     const role = await Role.create({ name });
     const permissions = await Permission.findAll({
       where: {
         id: ids,
       },
     });
+    if (!permissions.length) {
+      return res
+        .status(404)
+        .json({ message: "Permissions with provided ids does not exist. " });
+    }
+
     await role.addPermission(permissions);
-    return res.json(role);
+    return res.status(200).json(role);
   } catch (err) {
     console.log(err);
     return res.status(500).json(err);
@@ -25,7 +37,7 @@ router.get("/", async (req, res) => {
     const roles = await Role.findAll({
       include: Permission,
     });
-    return res.json(roles);
+    return res.status(200).json(roles);
   } catch (err) {
     console.log(err);
     return res.status(500).json(err);
@@ -56,7 +68,16 @@ router.put("/:id", async (req, res) => {
   const { name, selectedPermissions } = req.body;
   try {
     const ids = selectedPermissions.map((permission) => permission.id);
+
+    const isExistingRole = await Role.findOne({ where: { name } });
+    if (isExistingRole) {
+      return res.status(400).json({
+        message: "Role with provided name already exists.",
+      });
+    }
+
     await Role.update({ name }, { where: { id } });
+
     const role = await Role.findOne({ where: { id } });
     const permissions = await Permission.findAll({ where: { id: ids } });
 
@@ -71,8 +92,13 @@ router.put("/:id", async (req, res) => {
 
 router.delete("/:name", async (req, res) => {
   const { name } = req.params;
-  console.log("delete come");
   try {
+    const role = await Role.findOne({ where: { name } });
+    if (!role) {
+      return res
+        .status(404)
+        .json({ message: "Role with provided name does not exist." });
+    }
     await Role.destroy({
       where: {
         name,
